+++ include/pregpu.h
@@ -6,12 +6,12 @@
 
 static int   *keysDevc;                                         // Keys on device
 static int   *rangeDevc;                                        // Ranges on device
-static float *sourceDevc;                                       // Sources on device
-static float *targetDevc;                                       // Targets on device
-__device__ __constant__ float constDevc[1];                     // Constants on device
+static double *sourceDevc;                                      // Sources on device
+static double *targetDevc;                                      // Targets on device
+__device__ __constant__ double constDevc[1];                    // Constants on device
 
 namespace {
-__device__ void cart2sph(float& r, float& theta, float& phi, float dx, float dy, float dz) {
+__device__ void cart2sph(double& r, double& theta, double& phi, double dx, double dy, double dz) {
   r = sqrtf(dx * dx + dy * dy + dz * dz)+EPS;
   theta = acosf(dz / r);
   if( fabs(dx) + fabs(dy) < EPS ) {
@@ -25,7 +25,7 @@ __device__ void cart2sph(float& r, float& theta, float& phi, float dx, float dy,
   }
 }
 
-__device__ void sph2cart(float r, float theta, float phi, float *spherical, float *cartesian) {
+__device__ void sph2cart(double r, double theta, double phi, double *spherical, double *cartesian) {
   cartesian[0] = sinf(theta) * cosf(phi) * spherical[0]
                + cosf(theta) * cosf(phi) / r * spherical[1]
                - sinf(phi) / r / sinf(theta) * spherical[2];
@@ -36,28 +36,28 @@ __device__ void sph2cart(float r, float theta, float phi, float *spherical, floa
                - sinf(theta) / r * spherical[1];
 }
 
-__device__ void evalMultipole(float *YnmShrd, float rho, float alpha, float *factShrd) {
-  float x = cosf(alpha);
-  float s = sqrtf(1 - x * x);
-  float fact = 1;
-  float pn = 1;
-  float rhom = 1;
+__device__ void evalMultipole(double *YnmShrd, double rho, double alpha, double *factShrd) {
+  double x = cosf(alpha);
+  double s = sqrtf(1 - x * x);
+  double fact = 1;
+  double pn = 1;
+  double rhom = 1;
   for( int m=0; m<P; ++m ){
-    float p = pn;
+    double p = pn;
     int npn = m * m + 2 * m;
     int nmn = m * m;
     YnmShrd[npn] = rhom * p / factShrd[2*m];
     YnmShrd[nmn] = YnmShrd[npn];
-    float p1 = p;
+    double p1 = p;
     p = x * (2 * m + 1) * p;
     rhom *= -rho;
-    float rhon = rhom;
+    double rhon = rhom;
     for( int n=m+1; n<P; ++n ){
       int npm = n * n + n + m;
       int nmm = n * n + n - m;
       YnmShrd[npm] = rhon * p / factShrd[n+m];
       YnmShrd[nmm] = YnmShrd[npm];
-      float p2 = p1;
+      double p2 = p1;
       p1 = p;
       p = (x * (2 * n + 1) * p1 - (n + m) * p2) / (n - m + 1);
       rhon *= -rho;
@@ -67,24 +67,24 @@ __device__ void evalMultipole(float *YnmShrd, float rho, float alpha, float *fac
   }
 }
 
-__device__ void evalLocal(float *YnmShrd, float rho, float alpha, float *factShrd) {
-  float x = cosf(alpha);
-  float s = sqrtf(1 - x * x);
-  float fact = 1;
-  float pn = 1;
-  float rhom = 1.0 / rho;
+__device__ void evalLocal(double *YnmShrd, double rho, double alpha, double *factShrd) {
+  double x = cosf(alpha);
+  double s = sqrtf(1 - x * x);
+  double fact = 1;
+  double pn = 1;
+  double rhom = 1.0 / rho;
   for( int m=0; m<2*P; ++m ){
-    float p = pn;
+    double p = pn;
     int i = m * (m + 1) /2 + m;
     YnmShrd[i] = rhom * p;
-    float p1 = p;
+    double p1 = p;
     p = x * (2 * m + 1) * p;
     rhom /= rho;
-    float rhon = rhom;
+    double rhon = rhom;
     for( int n=m+1; n<2*P; ++n ){
       i = n * (n + 1) / 2 + m;
       YnmShrd[i] = rhon * p * factShrd[n-m];
-      float p2 = p1;
+      double p2 = p1;
       p1 = p;
       p = (x * (2 * n + 1) * p1 - (n + m) * p2) / (n - m + 1);
       rhon /= rho;
